LPS-X Move WorkflowTaskManager to portal-workflow-api and apply to usages
