+++ modules/apps/commerce/commerce-frontend-impl/build.gradle
@@ -24,6 +24,7 @@ dependencies {
 	compileOnly project(":apps:commerce:commerce-wish-list-api")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib-clay")
+	compileOnly project(":apps:portal-workflow:portal-workflow-api")
 	compileOnly project(":core:osgi-service-tracker-collections")
 	compileOnly project(":core:petra:petra-string")
 }
