+++ modules/apps/headless/headless-admin-workflow/headless-admin-workflow-impl/src/main/java/com/liferay/headless/admin/workflow/internal/resource/v1_0/TransitionResourceImpl.java
@@ -19,9 +19,9 @@ import com.liferay.headless.admin.workflow.internal.dto.v1_0.util.TransitionUtil
 import com.liferay.headless.admin.workflow.resource.v1_0.TransitionResource;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.workflow.WorkflowInstanceManager;
-import com.liferay.portal.kernel.workflow.WorkflowTaskManager;
 import com.liferay.portal.vulcan.pagination.Page;
 import com.liferay.portal.vulcan.pagination.Pagination;
+import com.liferay.portal.workflow.manager.WorkflowTaskManager;
 
 import java.util.List;
