+++ modules/apps/headless/headless-admin-workflow/headless-admin-workflow-test/build.gradle
@@ -8,5 +8,6 @@ dependencies {
 	testIntegrationCompile project(":apps:headless:headless-admin-workflow:headless-admin-workflow-client")
 	testIntegrationCompile project(":apps:portal-odata:portal-odata-api")
 	testIntegrationCompile project(":apps:portal-vulcan:portal-vulcan-api")
+	testIntegrationCompile project(":apps:portal-workflow:portal-workflow-api")
 	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
 }
