+++ modules/apps/layout/layout-test/build.gradle
@@ -30,6 +30,7 @@ dependencies {
 	testIntegrationCompile project(":apps:portal-search:portal-search-api")
 	testIntegrationCompile project(":apps:portal-search:portal-search-test-util")
 	testIntegrationCompile project(":apps:portal-vulcan:portal-vulcan-api")
+	testIntegrationCompile project(":apps:portal-workflow:portal-workflow-api")
 	testIntegrationCompile project(":apps:portal:portal-props-test-util")
 	testIntegrationCompile project(":apps:ratings:ratings-test-util")
 	testIntegrationCompile project(":apps:segments:segments-api")
