+++ modules/apps/object/object-test/build.gradle
@@ -21,5 +21,6 @@ dependencies {
 	testIntegrationCompile project(":apps:object:object-rest-api")
 	testIntegrationCompile project(":apps:object:object-scripting-api")
 	testIntegrationCompile project(":apps:portal-vulcan:portal-vulcan-api")
+	testIntegrationCompile project(":apps:portal-workflow:portal-workflow-api")
 	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
 }
